requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
azure-ai-documentintelligence==1.0.0b4
azure-core==1.32.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
black==23.3.0
boto3==1.28.65
botocore==1.31.85
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.20
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.1
cryptography==44.0.0
cssselect==1.2.0
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.15
distro==1.9.0
duckduckgo_search==6.3.7
durationpy==0.9
emoji==2.14.0
et_xmlfile==2.0.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
fastapi==0.115.5
filelock==3.16.1
filetype==1.2.0
Flask==2.1.1
flatbuffers==24.3.25
fonttools==4.55.3
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.13.1
grpcio==1.68.1
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
InstructorEmbedding==1.0.1
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kiwisolver==1.4.8
kubernetes==31.0.0
langchain==0.3.9
langchain-community==0.3.8
langchain-core==0.3.21
langchain-text-splitters==0.3.2
langdetect==1.0.9
langsmith==0.1.147
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.7.1
mdurl==0.1.2
mistralai==1.2.5
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.24.2
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.20.1
opencv-contrib-python==4.8.0.74
opencv-python==4.8.0.74
openpyxl==3.1.5
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
orjson==3.10.12
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parsel==1.9.1
pathspec==0.11.1
pdf2image==1.17.0
peewee==3.17.8
phidata==2.7.4
pillow==11.0.0
platformdirs==3.2.0
pluggy==1.5.0
posthog==3.7.4
primp==0.9.1
propcache==0.2.1
protobuf==5.29.0
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
pypdf==5.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pytesseract==0.3.13
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-oxmsg==0.0.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
s3transfer==0.7.0
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.15.0
sentence-transformers==3.3.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.41.3
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.3
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.67.0
transformers==4.46.2
typer==0.14.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.8
unstructured-client==0.28.1
urllib3==2.0.7
uvicorn==0.32.1
uvloop==0.21.0
w3lib==2.2.1
watchfiles==1.0.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==2.1.1
wrapt==1.17.0
yarl==1.18.3
yfinance==0.2.50
zipp==3.21.0